requirements.txt
apache-airflow-providers-google
pyarrow
kaggle
